server/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
anthropic>=0.49.0
python-dotenv==1.0.1
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.149.0
google-cloud-firestore>=2.16.0
google-ads>=25.0.0
google-generativeai>=0.8.0
openai>=1.40.0
